Move ServiceWait to the service package.
